Basic search funtionality in.
Slight refactoring of Installer and RemoteInstaller


git-svn-id: svn+ssh://rubyforge.org/var/svn/rubygems/trunk@105 3d4018f9-ac1a-0410-99e9-8a154d859a19

+++ rubygems/lib/rubygems/installer.rb
@@ -55,8 +55,9 @@ module Gem
        end
        
        write_spec(format.spec, File.join(to_dir, "specifications"))
-       
-       FileUtils.cp(@gem, File.join(to_dir, "cache"))
+       unless(File.exist?(File.join(File.join(to_dir, "cache"),@gem.split(/\//).pop))) 
+         FileUtils.cp(@gem, File.join(to_dir, "cache"))
+       end
        
        puts "Successfully installed #{format.spec.name} version #{format.spec.version}"
        format.spec.loaded_from = File.join(to_dir, 'specifications', format.spec.full_name+".gemspec")
